Changed rating level to a float.  Added a lot of test data in terms of new users, ratings, and subscriptions.  Also added an admin account which manages 3 food trucks which every user is subscribed to automatically (all users, customers only, and owners only).

+++ src/main/java/foodtruckfinder/site/common/External/Rating.java
@@ -5,19 +5,19 @@ import java.time.LocalDateTime;
 public class Rating {
     private String message;
     private LocalDateTime date;
-    private int rating;
+    private Float rating;
     private Long user_ID;
     private Long truck_ID;
 
     public String getMessage(){return this.message;}
     public LocalDateTime getDate() {return this.date;}
-    public int getRating() {return this.rating;}
+    public Float getRating() {return this.rating;}
     public Long getUser() {return this.user_ID;}
     public Long getTruck() {return this.truck_ID;}
 
     public void setMessage(String message){this.message = message;}
     public void setDate(LocalDateTime time){this.date = time;}
-    public void setRating(int rated){this.rating = rated;}
+    public void setRating(Float rated){this.rating = rated;}
     public void setUser(Long id){this.user_ID = id;}
     public void setTruck(Long id){this.truck_ID = id;}
 }
